Move KeyAddrEventQueue class from Qukeys to Kaleidoscope core
It's now being used by more than one plugin, and is likely to get used by at
least a third, if not more.
